Lorenz with Sindy and DMD
when I model the Lorenz problem with Sindy, it works well, but when I use DMD, it produces complex numbers as coefficients! The point is it looks normal and even the structure of DEs are true, the only problem is in the coefficients. The basis is set to be the same in both algorithms.